# perf(core): reduce assumption checking in Mul(x,y)