@oscarbenjamin 2023-03-09
Note that the effect of this might not show up in the benchmark CI job because the benchmarks don't account for cacheing and all assumptions queries on an expression are cached on the expression:
```python
In [1]: e = exp(2)

In [2]: e._assumptions
Out[2]: {}

In [3]: e.is_positive
Out[3]: True

In [4]: e._assumptions
Out[4]: 
{'extended_positive': True,
 'negative': False,
 'zero': False,
 'extended_real': True,
 'nonpositive': False,
 'extended_negative': False,
 'extended_nonzero': True,
 'extended_nonnegative': True,
 'imaginary': False,
 'extended_nonpositive': False,
 'commutative': True,
 'finite': True,
 'infinite': False,
 'nonnegative': True,
 'hermitian': True,
 'complex': True,
 'real': True,
 'nonzero': True,
 'positive': True}
```