@smichr 2023-03-09
That might have been an automated change since it used to be `if a and a.is_Rational` and was changed to `if not a.is_zero and a.is_Rational`.